+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/dsl/instrumentation/ActivityLifecycleConfiguration.kt
@@ -18,25 +18,26 @@ import io.opentelemetry.api.trace.Tracer
 @OpenTelemetryDslMarker
 class ActivityLifecycleConfiguration internal constructor(
     private val config: OtelRumConfig,
-) : ScreenLifecycleConfigurable,
-    CanBeEnabledAndDisabled {
-    private val activityLifecycleInstrumentation: ActivityLifecycleInstrumentation by lazy {
-        AndroidInstrumentationLoader.getInstrumentation(ActivityLifecycleInstrumentation::class.java)
+    private val instrumentationLoader: AndroidInstrumentationLoader,
+) : ScreenLifecycleConfigurable, CanBeEnabledAndDisabled {
+
+    private val activityLifecycleInstrumentation: ActivityLifecycleInstrumentation? by lazy {
+        instrumentationLoader.getByType(ActivityLifecycleInstrumentation::class.java)
     }
 
     override fun tracerCustomizer(value: (Tracer) -> Tracer) {
-        activityLifecycleInstrumentation.setTracerCustomizer(value)
+        activityLifecycleInstrumentation?.setTracerCustomizer(value)
     }
 
     override fun screenNameExtractor(value: ScreenNameExtractor) {
-        activityLifecycleInstrumentation.setScreenNameExtractor(value)
+        activityLifecycleInstrumentation?.setScreenNameExtractor(value)
     }
 
     override fun enabled(enabled: Boolean) {
         if (enabled) {
-            config.allowInstrumentation(activityLifecycleInstrumentation.name)
+            activityLifecycleInstrumentation?.name?.let { config.allowInstrumentation(it) }
         } else {
-            config.suppressInstrumentation(activityLifecycleInstrumentation.name)
+            activityLifecycleInstrumentation?.name?.let { config.suppressInstrumentation(it) }
         }
     }
 }