+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/dsl/instrumentation/FragmentLifecycleConfiguration.kt
@@ -18,25 +18,26 @@ import io.opentelemetry.api.trace.Tracer
 @OpenTelemetryDslMarker
 class FragmentLifecycleConfiguration internal constructor(
     private val config: OtelRumConfig,
-) : ScreenLifecycleConfigurable,
-    CanBeEnabledAndDisabled {
-    private val fragmentLifecycleInstrumentation: FragmentLifecycleInstrumentation by lazy {
-        AndroidInstrumentationLoader.getInstrumentation(FragmentLifecycleInstrumentation::class.java)
+    private val instrumentationLoader: AndroidInstrumentationLoader,
+) : ScreenLifecycleConfigurable, CanBeEnabledAndDisabled {
+
+    private val fragmentLifecycleInstrumentation: FragmentLifecycleInstrumentation? by lazy {
+        instrumentationLoader.getByType(FragmentLifecycleInstrumentation::class.java)
     }
 
     override fun tracerCustomizer(value: (Tracer) -> Tracer) {
-        fragmentLifecycleInstrumentation.setTracerCustomizer(value)
+        fragmentLifecycleInstrumentation?.setTracerCustomizer(value)
     }
 
     override fun screenNameExtractor(value: ScreenNameExtractor) {
-        fragmentLifecycleInstrumentation.setScreenNameExtractor(value)
+        fragmentLifecycleInstrumentation?.setScreenNameExtractor(value)
     }
 
     override fun enabled(enabled: Boolean) {
         if (enabled) {
-            config.allowInstrumentation(fragmentLifecycleInstrumentation.name)
+            fragmentLifecycleInstrumentation?.name?.let { config.allowInstrumentation(it) }
         } else {
-            config.suppressInstrumentation(fragmentLifecycleInstrumentation.name)
+            fragmentLifecycleInstrumentation?.name?.let { config.suppressInstrumentation(it) }
         }
     }
 }