+++ android-agent/src/test/kotlin/io/opentelemetry/android/agent/dsl/ClockConfigTest.kt
@@ -8,25 +8,33 @@ package io.opentelemetry.android.agent.dsl
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.opentelemetry.android.OtelAndroidClock
 import io.opentelemetry.android.agent.FakeClock
+import io.opentelemetry.android.agent.FakeInstrumentationLoader
 import io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.assertThat
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
 class ClockConfigTest {
+
+    private lateinit var otelConfig: OpenTelemetryConfiguration
+
+    @Before
+    fun setUp() {
+        otelConfig = OpenTelemetryConfiguration(instrumentationLoader = FakeInstrumentationLoader())
+    }
+
     @Test
     fun testDefaults() {
-        val otelConfig = OpenTelemetryConfiguration()
         assertThat(otelConfig.clock).isInstanceOf(OtelAndroidClock::class.java)
     }
 
     @Test
     fun testOverride() {
         val fakeClock = FakeClock()
-        val otelConfig =
-            OpenTelemetryConfiguration().apply {
-                clock = fakeClock
-            }
+        val otelConfig = otelConfig.apply {
+            clock = fakeClock
+        }
         assertThat(otelConfig.clock).isSameAs(fakeClock)
     }
 }